# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MemeHub API - 9GAG Clone
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get posts feed
      parameters:
        - name: section
          in: query
          schema:
            type: string
            enum: [hot, trending, fresh, top]
            default: hot
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsListResponse"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /comments/{id}/vote:
    post:
      operationId: voteComment
      tags: [votes]
      summary: Vote on a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResponse"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsListResponse"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersListResponse"
...
  /admin/posts:
    get:
      operationId: getAdminPosts
      tags: [admin]
      summary: Get all posts (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, removed]
            default: all
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPostsListResponse"
...
  /admin/tags:
    post:
      operationId: createTag
      tags: [admin]
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
...
  /admin/comments:
    get:
      operationId: getAdminComments
      tags: [admin]
      summary: Get all comments (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsListResponse"
  /search:
    get:
      operationId: searchPosts
      tags: [posts]
      summary: Search posts
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsListResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required:
        - user
        - token
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin]
        isBanned:
          type: boolean
        followersCount:
          type: number
        followingCount:
          type: number
        postsCount:
          type: number
        totalPoints:
          type: number
        isFollowing:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - role
        - isBanned
        - followersCount
        - followingCount
        - postsCount
        - totalPoints
        - isFollowing
        - createdAt
    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        gifUrl:
          type: string
          nullable: true
        type:
          type: string
          enum: [image, gif, video]
        upvotes:
          type: number
        downvotes:
          type: number
        points:
          type: number
        commentsCount:
          type: number
        viewsCount:
          type: number
        status:
          type: string
          enum: [pending, approved, removed]
        userVote:
          type: string
          enum: [up, down]
          nullable: true
        isSaved:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        author:
          $ref: "#/components/schemas/PostAuthor"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - imageUrl
        - type
        - upvotes
        - downvotes
        - points
        - commentsCount
        - viewsCount
        - status
        - isSaved
        - tags
        - author
        - createdAt
    PostAuthor:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
          nullable: true
      required:
        - id
        - username
    PostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
        hasMore:
          type: boolean
      required:
        - posts
        - total
        - page
        - totalPages
        - hasMore
    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        upvotes:
          type: number
        downvotes:
          type: number
        userVote:
          type: string
          enum: [up, down]
          nullable: true
        author:
          $ref: "#/components/schemas/PostAuthor"
        parentId:
          type: string
          nullable: true
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - upvotes
        - downvotes
        - author
        - createdAt
    CommentsListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - comments
        - total
        - page
        - totalPages
...
    VoteRequest:
      type: object
      properties:
        direction:
          type: string
          enum: [up, down]
      required:
        - direction
    VoteResponse:
      type: object
      properties:
        upvotes:
          type: number
        downvotes:
          type: number
        points:
          type: number
        userVote:
          type: string
          enum: [up, down]
          nullable: true
      required:
        - upvotes
        - downvotes
        - points
...
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        postsCount:
          type: number
        color:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
        - postsCount
...
    CreateTagRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
      required:
        - name
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        imageUrl:
          type: string
        type:
          type: string
          enum: [image, gif, video]
        tagIds:
          type: array
          items:
            type: string
      required:
        - title
        - imageUrl
        - type
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [upvote, comment, reply, follow, mention]
        message:
          type: string
        isRead:
          type: boolean
        postId:
          type: string
          nullable: true
        fromUser:
          $ref: "#/components/schemas/PostAuthor"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - message
        - isRead
        - fromUser
        - createdAt
    NotificationsListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: number
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - notifications
        - unreadCount
        - total
        - page
        - totalPages
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin]
        isBanned:
          type: boolean
        postsCount:
          type: number
        commentsCount:
          type: number
        totalPoints:
          type: number
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
        - email
        - role
        - isBanned
        - postsCount
        - commentsCount
        - totalPoints
        - createdAt
    AdminUsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - users
        - total
        - page
        - totalPages
    AdminPostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - posts
        - total
        - page
        - totalPages
...